fix pixel type issue when displaying RGB image with projected point cloud --- image in display was completely white

+++ python/project_laser_into_camera.py
@@ -72,7 +72,7 @@ image = load_image(image_path, model)
 
 uv, depth = model.project(pointcloud, image.shape)
 
-plt.imshow(image)
+plt.imshow(image.astype("int64"))
 plt.hold(True)
 plt.scatter(np.ravel(uv[0, :]), np.ravel(uv[1, :]), s=2, c=depth, edgecolors='none', cmap='jet')
 plt.xlim(0, image.shape[1])
